+++ code/classeine-lib/clsn/ui/defaults/ButtonDefaults.cpp
@@ -1,15 +1,9 @@
 #include "ButtonDefaults.h"
 
-#include "clsn/ui/renderers/ButtonRenderer.h"
 
 namespace clsn::ui::defaults::ButtonDefaults
 {
-    using namespace clsn::draw;
-
-    using namespace clsn::ui::renderers;
-
-    void populateDefaults(clsn::ui::UIManagerDefaults& defaults)
+    void populateDefaults(clsn::ui::UIManagerDefaults& )
     {
-        defaults.set("Button", "renderer", makeLazyRenderer<ButtonRenderer>());
     }
 }