+++ code/classeine-lib/clsn/ui/defaults/CheckBoxDefaults.cpp
@@ -1,15 +1,9 @@
 #include "CheckBoxDefaults.h"
 
-#include "clsn/ui/renderers/CheckBoxRenderer.h"
 
 namespace clsn::ui::defaults::CheckBoxDefaults
 {
-    using namespace clsn::draw;
-
-    using namespace clsn::ui::renderers;
-
-    void populateDefaults(clsn::ui::UIManagerDefaults& defaults)
+    void populateDefaults(clsn::ui::UIManagerDefaults& )
     {
-        defaults.set("CheckBox", "renderer", makeLazyRenderer<CheckBoxRenderer>());
     }
 }