+++ code/classeine-lib/clsn/ui/defaults/DefaultUISkin.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "clsn/ui/UISkin.h"
+
+namespace clsn::ui::defaults
+{
+    class DefaultUISkin : public UISkin
+    {
+    public:
+        DefaultUISkin();
+    };
+}