+++ code/classeine-lib/clsn/ui/defaults/HBoxContainerDefaults.cpp
@@ -1,15 +1,9 @@
 #include "HBoxContainerDefaults.h"
 
-#include "clsn/ui/renderers/HBoxContainerRenderer.h"
 
 namespace clsn::ui::defaults::HBoxContainerDefaults
 {
-    using namespace clsn::draw;
-
-    using namespace clsn::ui::renderers;
-
-    void populateDefaults(clsn::ui::UIManagerDefaults& defaults)
+    void populateDefaults(clsn::ui::UIManagerDefaults& )
     {
-        defaults.set("HBoxContainer", "renderer", makeLazyRenderer<HBoxContainerRenderer>());
     }
 }