+++ code/classeine-lib/clsn/ui/defaults/LabelDefaults.cpp
@@ -1,15 +1,9 @@
 #include "LabelDefaults.h"
 
-#include "clsn/ui/renderers/LabelRenderer.h"
 
 namespace clsn::ui::defaults::LabelDefaults
 {
-    using namespace clsn::draw;
-
-    using namespace clsn::ui::renderers;
-
-    void populateDefaults(clsn::ui::UIManagerDefaults& defaults)
+    void populateDefaults(clsn::ui::UIManagerDefaults& )
     {
-        defaults.set("Label", "renderer", makeLazyRenderer<LabelRenderer>());
     }
 }