+++ code/classeine-lib/clsn/ui/defaults/RadioButtonDefaults.cpp
@@ -1,15 +1,9 @@
 #include "RadioButtonDefaults.h"
 
-#include "clsn/ui/renderers/RadioButtonRenderer.h"
 
 namespace clsn::ui::defaults::RadioButtonDefaults
 {
-    using namespace clsn::draw;
-
-    using namespace clsn::ui::renderers;
-
-    void populateDefaults(clsn::ui::UIManagerDefaults& defaults)
+    void populateDefaults(clsn::ui::UIManagerDefaults& )
     {
-        defaults.set("RadioButton", "renderer", makeLazyRenderer<RadioButtonRenderer>());
     }
 }