+++ code/classeine-lib/clsn/ui/defaults/ToggleButtonDefaults.cpp
@@ -1,15 +1,9 @@
 #include "ToggleButtonDefaults.h"
 
-#include "clsn/ui/renderers/ToggleButtonRenderer.h"
 
 namespace clsn::ui::defaults::ToggleButtonDefaults
 {
-    using namespace clsn::draw;
-
-    using namespace clsn::ui::renderers;
-
-    void populateDefaults(clsn::ui::UIManagerDefaults& defaults)
+    void populateDefaults(clsn::ui::UIManagerDefaults& )
     {
-        defaults.set("ToggleButton", "renderer", makeLazyRenderer<ToggleButtonRenderer>());
     }
 }