+++ code/classeine-lib/clsn/ui/defaults/VBoxContainerDefaults.cpp
@@ -1,15 +1,8 @@
 #include "VBoxContainerDefaults.h"
 
-#include "clsn/ui/renderers/VBoxContainerRenderer.h"
-
 namespace clsn::ui::defaults::VBoxContainerDefaults
 {
-    using namespace clsn::draw;
-
-    using namespace clsn::ui::renderers;
-
-    void populateDefaults(clsn::ui::UIManagerDefaults& defaults)
+    void populateDefaults(clsn::ui::UIManagerDefaults& )
     {
-        defaults.set("VBoxContainer", "renderer", makeLazyRenderer<VBoxContainerRenderer>());
     }
 }