+++ code/classeine-lib/clsn/ui/defaults/XYContainerDefaults.cpp
@@ -1,15 +1,8 @@
 #include "XYContainerDefaults.h"
 
-#include "clsn/ui/renderers/XYContainerRenderer.h"
-
 namespace clsn::ui::defaults::XYContainerDefaults
 {
-    using namespace clsn::draw;
-
-    using namespace clsn::ui::renderers;
-
-    void populateDefaults(clsn::ui::UIManagerDefaults& defaults)
+    void populateDefaults(clsn::ui::UIManagerDefaults& )
     {
-        defaults.set("XYContainer", "renderer", makeLazyRenderer<XYContainerRenderer>());
     }
 }